# docs/swagger.yaml
definitions:
...
  mengri-flow_internal_app_dto.ClusterDetailResponse:
    properties:
      activeFlows:
        type: integer
      createdAt:
        type: string
      description:
        type: string
      environmentId:
        type: string
      etcdEndpoints:
        type: string
      etcdUsername:
        type: string
      executorCount:
        type: integer
      id:
        type: string
      name:
        type: string
      nodeCount:
        type: integer
      status:
        type: string
      updatedAt:
        type: string
    type: object
  mengri-flow_internal_app_dto.ClusterResponse:
    properties:
      createdAt:
        type: string
      description:
        type: string
      environmentId:
        type: string
      etcdEndpoints:
        type: string
      etcdUsername:
        type: string
      id:
        type: string
      name:
        type: string
      nodeCount:
        type: integer
      status:
        type: string
      updatedAt:
        type: string
    type: object
...
  mengri-flow_internal_app_dto.CreateClusterRequest:
    properties:
      description:
        maxLength: 500
        type: string
      environmentId:
        type: string
      etcdEndpoints:
        type: string
      etcdPassword:
        type: string
      etcdUsername:
        type: string
      name:
        maxLength: 100
        type: string
    required:
    - environmentId
    - etcdEndpoints
    - name
    type: object
  mengri-flow_internal_app_dto.CreateEnvironmentRequest:
    properties:
      color:
        type: string
      description:
        maxLength: 200
        type: string
      key:
        maxLength: 20
        type: string
      name:
        maxLength: 50
        type: string
    required:
    - key
    - name
    type: object
...
  mengri-flow_internal_app_dto.EnvironmentResponse:
    properties:
      clusterCount:
        type: integer
      color:
        type: string
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      key:
        type: string
      name:
        type: string
      updatedAt:
        type: string
    type: object
...
  mengri-flow_internal_app_dto.ListClustersResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/mengri-flow_internal_app_dto.ClusterResponse'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  mengri-flow_internal_app_dto.ListEnvironmentsResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/mengri-flow_internal_app_dto.EnvironmentResponse'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  mengri-flow_internal_app_dto.TestEtcdConnectionRequest:
    properties:
      endpoints:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - endpoints
    type: object
...
  mengri-flow_internal_app_dto.UpdateClusterRequest:
    properties:
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 100
        type: string
    type: object
  mengri-flow_internal_app_dto.UpdateEnvironmentRequest:
    properties:
      color:
        type: string
      description:
        maxLength: 200
        type: string
      name:
        maxLength: 50
        type: string
    type: object
...
  mengri-flow_pkg_response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /clusters:
    get:
      consumes:
      - application/json
      parameters:
      - description: 环境ID
        in: query
        name: environmentId
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/mengri-flow_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/mengri-flow_internal_app_dto.ListClustersResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 获取集群列表
      tags:
      - Cluster
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建集群请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/mengri-flow_internal_app_dto.CreateClusterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/mengri-flow_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/mengri-flow_internal_app_dto.ClusterResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 创建集群
      tags:
      - Cluster
  /clusters/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 集群ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 删除集群
      tags:
      - Cluster
    get:
      consumes:
      - application/json
      parameters:
      - description: 集群ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/mengri-flow_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/mengri-flow_internal_app_dto.ClusterDetailResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 获取集群详情
      tags:
      - Cluster
    put:
      consumes:
      - application/json
      parameters:
      - description: 集群ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新集群请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/mengri-flow_internal_app_dto.UpdateClusterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/mengri-flow_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/mengri-flow_internal_app_dto.ClusterResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 更新集群
      tags:
      - Cluster
  /clusters/{id}/test-etcd:
    post:
      consumes:
      - application/json
      parameters:
      - description: 集群ID
        in: path
        name: id
        required: true
        type: string
      - description: 测试连接请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/mengri-flow_internal_app_dto.TestEtcdConnectionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 测试etcd连接
      tags:
      - Cluster
  /environments:
    get:
      consumes:
      - application/json
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/mengri-flow_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/mengri-flow_internal_app_dto.ListEnvironmentsResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 获取环境列表
      tags:
      - Environment
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建环境请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/mengri-flow_internal_app_dto.CreateEnvironmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/mengri-flow_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/mengri-flow_internal_app_dto.EnvironmentResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 创建环境
      tags:
      - Environment
  /environments/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 环境ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 删除环境
      tags:
      - Environment
    get:
      consumes:
      - application/json
      parameters:
      - description: 环境ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/mengri-flow_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/mengri-flow_internal_app_dto.EnvironmentResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 获取环境详情
      tags:
      - Environment
    put:
      consumes:
      - application/json
      parameters:
      - description: 环境ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新环境请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/mengri-flow_internal_app_dto.UpdateEnvironmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/mengri-flow_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/mengri-flow_internal_app_dto.EnvironmentResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 更新环境
      tags:
      - Environment
...
swagger: "2.0"